data/data_fetcher.py
# data/data_fetcher.py

import logging

class DataFetcher:
    def __init__(self, stock_symbol):
        self.stock_symbol = stock_symbol

    def get_historical_data(self):
        logging.info(f"Fetching historical data for {self.stock_symbol}")
        # Stub: Replace with actual API calls to fetch historical stock data.
        # Here we simulate with mock data for testing purposes.
        historical_data = [100, 102, 101, 105, 110]  # Mock data
        return historical_data

# Debugging Instructions:
# - Replace the stub with actual API calls to fetch historical data.
# - Test with different stock symbols to ensure the data fetched is accurate.
# - Use the test script to validate the integrity and performance of this function.